Reworked new qualifications details page.
