+++ src/SFA.DAS.AODP.Application/Queries/Qualifications/GetProcessStatusesQuery.cs
@@ -4,5 +4,4 @@ namespace SFA.DAS.AODP.Application.Queries.Qualifications;
 
 public class GetProcessStatusesQuery : IRequest<BaseMediatrResponse<GetProcessStatusesQueryResponse>>
 {
-    public string QualificationReference { get; set; } = string.Empty;
 }
