+++ src/SFA.DAS.AODP.Domain/Qualifications/Requests/GetChangedQualificationCSVExportApiRequest.cs
@@ -1,9 +1,8 @@
 ﻿using SFA.DAS.AODP.Domain.Interfaces;
 
-namespace SFA.DAS.AODP.Domain.Qualifications.Requests
+namespace SFA.DAS.AODP.Domain.Qualifications.Requests;
+
+public class GetChangedQualificationCsvExportApiRequest : IGetApiRequest
 {
-    public class GetProcessStatusesApiRequest : IGetApiRequest
-    {
-        public string GetUrl => "api/qualifications/processstatuses";
-    }
+    public string GetUrl => "api/qualifications/export?status=changed";
 }
