+++ src/SFA.DAS.AODP.Domain/Qualifications/Requests/GetProcessStatusesApiRequest.cs
@@ -1,9 +1,9 @@
 ﻿using SFA.DAS.AODP.Domain.Interfaces;
 
-namespace SFA.DAS.AODP.Domain.Qualifications.Requests
+namespace SFA.DAS.AODP.Domain.Qualifications.Requests;
+
+
+public class GetProcessStatusesApiRequest : IGetApiRequest
 {
-    public class GetChangedQualificationCsvExportApiRequest : IGetApiRequest
-    {
-        public string GetUrl => "api/qualifications/export?status=changed";
-    }
+    public string GetUrl => "api/qualifications/processstatuses";
 }
