+++ src/SFA.DAS.AODP.Web/wwwroot/stylesheets/application.css
@@ -22,7 +22,7 @@
     position: relative;
     top: -8px;
     margin: 15px 0;
-    padding: 0 0 15px 18px;
+    padding: 0 0 0 18px;
 }
 
 .aodp-timeline__item::before {
